32. Личный проект: дополнительная функциональность Правки по критериям (порядок методов и тд)
