+++ src/constants.js
@@ -18,6 +18,7 @@ const TITLE_MESSAGE = {
 
 const ERROR_MESSAGE = {
   SYNC_FAILED: `Sync data failed`,
+  OFFLINE_LOGIC: `offline logic is not implemented`,
 };
 
 const CardsOther = {
