+++ src/utils/render.js
@@ -1,7 +1,6 @@
-import * as util from "./common.js";
-
 const TITLE_ELEMENT_CLASS = `films-list__title`;
 const TITLE_ERROR_CLASS = `${TITLE_ELEMENT_CLASS}--error`;
+const HIDDEN_CLASS = `visually-hidden`;
 
 const renderElement = (container, component, method = `append`) => {
   if (!container) {
@@ -23,14 +22,22 @@ const createElement = (template, isWrapped) => {
   return isWrapped ? element : element.firstChild;
 };
 
+const showElement = (element) => {
+  element.classList.remove(HIDDEN_CLASS);
+};
+
+const hideElement = (element) => {
+  element.classList.add(HIDDEN_CLASS);
+};
+
 const showTitle = (message, element = document.querySelector(`.${TITLE_ELEMENT_CLASS}`)) => {
-  util.showElement(element);
+  showElement(element);
   element.textContent = message;
 };
 
 const hideTitle = (element = document.querySelector(`.${TITLE_ELEMENT_CLASS}`)) => {
   element.classList.remove(TITLE_ERROR_CLASS);
-  util.hideElement(element);
+  hideElement(element);
 };
 
 const showError = (message) => {
@@ -41,6 +48,8 @@ const showError = (message) => {
 export {
   renderElement,
   createElement,
+  showElement,
+  hideElement,
   showTitle,
   hideTitle,
   showError,
